requirements.txt
cached-property==1.4.3
Django==2.0.6
django-cors-headers==2.3.0
django-url-filter==0.3.5
djangorestframework==3.8.2
enum-compat==0.0.2
Pillow==5.1.0
pytz==2018.4
six==1.11.0
psycopg2==2.7.2